clean up ui util
